chore: rename nms to api
