+++ ui/package.json
@@ -1,6 +1,6 @@
 {
-  "name": "sdcore-nms",
-  "version": "0.2.0",
+  "name": "ella-core-ui",
+  "version": "0.0.0",
   "private": true,
   "scripts": {
     "dev": "next dev -p 3005",
